requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.10.0
numpy==1.23.5
pillow==9.5.0
gunicorn==20.1.0
werkzeug==2.2.3